# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bloum Cash API - Fintech PWA for Togo
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  /**
     * @minLength 6
     * @maxLength 6
     */
  pin: string;
}

export interface RegisterInput {
  fullName: string;
  email: string;
  /**
     * @minLength 6
     * @maxLength 6
     */
  pin: string;
}

export interface ForgotPinInput {
  email: string;
}

export interface User {
  id: string;
  fullName: string;
  email: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface MessageResponse {
  message: string;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  incoming: 'incoming',
  outgoing: 'outgoing',
} as const;

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  success: 'success',
  failed: 'failed',
  pending: 'pending',
} as const;

export type TransactionOperator = typeof TransactionOperator[keyof typeof TransactionOperator];


export const TransactionOperator = {
  tmoney: 'tmoney',
  moov: 'moov',
} as const;

export interface Transaction {
  id: string;
  type: TransactionType;
  title: string;
  amount: number;
  date: string;
  time?: string;
  status: TransactionStatus;
  operator: TransactionOperator;
  reference?: string;
  /** @nullable */
  fromPhone?: string | null;
  /** @nullable */
  toPhone?: string | null;
  /** @nullable */
  fees?: number | null;
  /** @nullable */
  description?: string | null;
}

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  incoming: 'incoming',
  outgoing: 'outgoing',
} as const;

export type TransactionInputOperator = typeof TransactionInputOperator[keyof typeof TransactionInputOperator];


export const TransactionInputOperator = {
  tmoney: 'tmoney',
  moov: 'moov',
} as const;

export interface TransactionInput {
  type: TransactionInputType;
  title: string;
  amount: number;
  operator: TransactionInputOperator;
  fromPhone?: string;
  toPhone?: string;
  description?: string;
}

export interface StatsSummary {
  totalAmount: number;
  incoming: number;
  outgoing: number;
  transactionCount: number;
  period?: string;
}

export interface ChartPoint {
  label: string;
  value: number;
}

export type QrInputOperator = typeof QrInputOperator[keyof typeof QrInputOperator];


export const QrInputOperator = {
  tmoney: 'tmoney',
  moov: 'moov',
} as const;

export interface QrInput {
  businessName: string;
  phone: string;
  operator: QrInputOperator;
  amount: number;
  description?: string;
}

export type QrCodeOperator = typeof QrCodeOperator[keyof typeof QrCodeOperator];


export const QrCodeOperator = {
  tmoney: 'tmoney',
  moov: 'moov',
} as const;

export interface QrCode {
  reference: string;
  businessName: string;
  phone: string;
  operator: QrCodeOperator;
  amount: number;
  qrData: string;
  /** @nullable */
  description?: string | null;
  createdAt?: string;
}

export type PaymentInputPayerOperator = typeof PaymentInputPayerOperator[keyof typeof PaymentInputPayerOperator];


export const PaymentInputPayerOperator = {
  tmoney: 'tmoney',
  moov: 'moov',
} as const;

export interface PaymentInput {
  payerPhone: string;
  payerOperator: PaymentInputPayerOperator;
}

export interface PaymentResult {
  success: boolean;
  message: string;
  reference: string;
  /** @nullable */
  transactionId?: string | null;
}

export type TransferInputFromOperator = typeof TransferInputFromOperator[keyof typeof TransferInputFromOperator];


export const TransferInputFromOperator = {
  tmoney: 'tmoney',
  moov: 'moov',
} as const;

export type TransferInputToOperator = typeof TransferInputToOperator[keyof typeof TransferInputToOperator];


export const TransferInputToOperator = {
  tmoney: 'tmoney',
  moov: 'moov',
} as const;

export interface TransferInput {
  fromOperator: TransferInputFromOperator;
  fromPhone: string;
  toOperator: TransferInputToOperator;
  toPhone: string;
  amount: number;
}

export interface TransferResult {
  success: boolean;
  message: string;
  reference: string;
  fees?: number;
  total?: number;
}

export type TransferFeesInputFromOperator = typeof TransferFeesInputFromOperator[keyof typeof TransferFeesInputFromOperator];


export const TransferFeesInputFromOperator = {
  tmoney: 'tmoney',
  moov: 'moov',
} as const;

export type TransferFeesInputToOperator = typeof TransferFeesInputToOperator[keyof typeof TransferFeesInputToOperator];


export const TransferFeesInputToOperator = {
  tmoney: 'tmoney',
  moov: 'moov',
} as const;

export interface TransferFeesInput {
  fromOperator: TransferFeesInputFromOperator;
  toOperator: TransferFeesInputToOperator;
  amount: number;
}

export interface TransferFees {
  amount: number;
  fees: number;
  total: number;
  estimatedTime: string;
}

export type ListTransactionsParams = {
filter?: string;
search?: string;
period?: string;
};

export type GetStatsSummaryParams = {
period?: GetStatsSummaryPeriod;
};

export type GetStatsSummaryPeriod = typeof GetStatsSummaryPeriod[keyof typeof GetStatsSummaryPeriod];


export const GetStatsSummaryPeriod = {
  today: 'today',
  week: 'week',
  month: 'month',
  year: 'year',
} as const;

export type GetStatsChartParams = {
period?: GetStatsChartPeriod;
};

export type GetStatsChartPeriod = typeof GetStatsChartPeriod[keyof typeof GetStatsChartPeriod];


export const GetStatsChartPeriod = {
  today: 'today',
  week: 'week',
  month: 'month',
  year: 'year',
} as const;