# lib/api-zod/src/generated/types/forgotPinInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bloum Cash API - Fintech PWA for Togo
 * OpenAPI spec version: 0.1.0
 */

export interface ForgotPinInput {
  email: string;
}