# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bloum Cash API - Fintech PWA for Togo
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  fullName: string;
  email: string;
  /**
     * @minLength 6
     * @maxLength 6
     */
  pin: string;
}