# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bloum Cash API - Fintech PWA for Togo
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalAmount: number;
  incoming: number;
  outgoing: number;
  transactionCount: number;
  period?: string;
}